cart with size and remove all cart items

+++ src/services/cart.ts
@@ -2,16 +2,18 @@ import { Request } from "express";
 import { UserTable, CartTable, ProductTable } from "../relation/index";
 import { validatePayload } from "../helper/payloadValidation";
 import { reduceStock } from "../utils/reduceStock";
+import { checkCartParameter, isValidSize } from "../utils/func";
 
 async function _createCartItemService(req: Request) {
   try {
     const {
       customer_id,
       product_id,
-      quantity,
-    }: { customer_id: number; product_id: number; quantity: number } = req.body;
+      size,
+      status
+    }: { customer_id: string; product_id: string; status: string; size: string } = req.body;
 
-    const requiredFields = ["customer_id", "product_id", "quantity"];
+    const requiredFields = ["customer_id", "product_id", "status", "size"];
     validatePayload(req.body, requiredFields);
 
     const user = await UserTable.findByPk(customer_id);
@@ -24,59 +26,84 @@ async function _createCartItemService(req: Request) {
       throw new Error("Product not found");
     }
 
-    if (product.stock < quantity) {
-      throw new Error("Insufficient quantity in stock");
+    if (!product.size.split(',').includes(size)) {
+      throw new Error("Size not available");
     }
 
-    if (quantity == 0) {
-      // If the quantity is zero, call the delete cart API
-      const response = await fetch(`http://localhost:4321/api/cart/${customer_id}/${product_id}`, {
-        method: 'DELETE',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-      });
-      const data = await response.json();
-      return { data : data.data , message: data.message , };
-    } else {
-      const existingCartItem: any = await CartTable.findOne({
-        where: {
-          customer_id,
-          product_id,
-        },
-      });
+    if (!isValidSize(size)) {
+      throw new Error("Invalid size");
+    }
 
-      let cartItem
+    if(!checkCartParameter(status)){
+      throw new Error("Status Parameter is invalid");      
+    }
+
+    const existingCartItem: any = await CartTable.findOne({
+      where: {
+        customer_id,
+        product_id,
+        size,
+      },
+    });
+
+    let cartItem;
+
+    if (status === 'add') {
       if (existingCartItem) {
-        // If the same product is already in the cart, increase the quantity
-        existingCartItem.quantity += Number(quantity);
+        existingCartItem.quantity += 1;
+        if (existingCartItem.quantity > product.stock) {
+          throw new Error("Insufficient stock");
+        }
         cartItem = await existingCartItem.save();
       } else {
-        // If the product is not in the cart, create a new cart item
+        if (product.stock < 1) {
+          throw new Error("Insufficient stock");
+        }
         cartItem = await CartTable.create({
           customer_id,
           product_id,
-          quantity,
+          quantity: 1,
+          size : size.toLocaleUpperCase(),
         });
       }
-      // Reduce the stock quantity in the product table //! for now reducing stock only when user buys the product
-      // await reduceStock(product, Number(quantity));
-      return {
-        data: cartItem,
-        message: "Item is added to the cart",
-      };
+    } else if (status === 'remove') {
+      if (!existingCartItem) {
+        return {
+          message: "Item not found",
+        };
+      }
+      existingCartItem.quantity -= 1;
+      if (existingCartItem.quantity == 0) {
+        await CartTable.destroy({
+          where: {
+            customer_id,
+            product_id,
+            size,
+          },
+        });
+        return {
+          message: "Item removed from the cart",
+        };
+      }
+      cartItem = await existingCartItem.save();
     }
 
+    return {
+      data: cartItem,
+      message: "Item " + (status === 'add' ? 'added to' : 'removed from') + " the cart",
+    };
   } catch (error) {
     console.error("Error in _createCartItemService:", error);
     throw error;
   }
 }
 
+
+
 async function _getCartListService(req: Request) {
   try {
     const { customer_id } = req.params;
-    if(!customer_id){
+    if (!customer_id) {
       throw new Error("Customer ID is required")
     }
     const customer = await UserTable.findByPk(customer_id);
@@ -99,19 +126,23 @@ async function _getCartListService(req: Request) {
 
 async function _removeProductFromCartService(req: Request) {
   try {
-    const { customer_id, product_id } = req.params;
-    if(!customer_id || !product_id){
-      throw new Error("Product ID and Customer ID is required")
+    const { customer_id, product_id, size } = req.params;
+    if (!customer_id || !product_id || !size) {
+      throw new Error("Product ID ,Customer ID ,size is required")
+    }
+    if (!isValidSize(size)) {
+      throw new Error("Invalid size");
     }
     const removedItem = await CartTable.destroy({
       where: {
         customer_id,
         product_id,
+        size : size.toLocaleUpperCase()
       },
     });
 
     if (!removedItem) {
-      throw new Error("Product or User not found")
+      throw new Error("User or product with this size not found")
     }
     return {
       data: removedItem,
@@ -122,5 +153,28 @@ async function _removeProductFromCartService(req: Request) {
     throw error;
   }
 }
+async function _removeAllProductFromCartService(req: Request) {
+  try {
+    const { customer_id } = req.params;
+    if (!customer_id) {
+      throw new Error("Customer ID is required")
+    }
+    const removedItem = await CartTable.destroy({
+      where: {
+        customer_id,
+      },
+    });
 
-export { _createCartItemService, _getCartListService, _removeProductFromCartService };
+    if (!removedItem) {
+      throw new Error("User has no item in cart")
+    }
+    return {
+      data: removedItem,
+      message: "All items are removed from cart",
+    };
+  } catch (error) {
+    console.error("Error in _removeAllProductFromCartService:", error);
+    throw error;
+  }
+}
+export { _createCartItemService, _getCartListService, _removeProductFromCartService , _removeAllProductFromCartService };
